Update README.md
In Lindgren et al. introduced PRoPHET (Probabilistic Routing Protocol using History of Encounters and Transitivity), a routing protocol that maintains the probability of reaching each node in the network. The protocol uses the history of previously encountered nodes and considers the transitive property inherent in human contacts to estimate the Delivery Predictability (DP) of a node, a metric that is used to predict the probability for a relay node to successfully deliver a message to its destination. In general, nodes that are more often encountered have a high DP value.
The PRoPHET protocol is based on the idea that if a node has visited a location or contacted another node frequently, the chances that this node to visit that location again in the near future are very high. Based on this, a node only forwards a message to an encountered relay node that has a higher DP value than its own DP value. This is meant to ensure that the message can only be forwarded to a more reliable relay node. A DP value is calculated based on three parameters: Updated metric, aging, and transitivity: 
Updated metric: whenever two nodes encounter each other, their DP is updated using Equation 1. Nodes that often encounter each other have high DP and they are most likely to be good message forwarders of one another.  

P (A, B) = P (A, B) old + (1 – P (A, B) old) × Pinit ………………………………………... (1) 

The P (A,B) is the delivery probability of node A to node B, P (A,B)old is the encounter probability of node A and B before the current encounter, and Pinit is an initialization constant. Whenever node A (that has a message for destination D) encounters a relay node B, both nodes A and B will exchange their DP values. Next, node A the compares DP value of P(A, D) against 𝑃 (B, D). If 𝑃 (B, D) is greater than P (A, D), the message is forwarded to node B; otherwise, it is not forwarded to node B. 
Aging: The basic idea is of aging is that, nodes that hardly encounter each other are unlikely to encounter each other in the nearest future; therefore, they are also unlikely to be good message forwarders of one another. Therefore, their DP is aged accordingly based on Equation 2.
P(A, ἱ) = P (A, ἱ) old × γT ….……………………………………..……… (2) 

Where P (A, ἱ) denotes the DP of nodes A, γ represents the aging constant (and its value is 0.998) and T is the number of time units elapsed since the last time P(A, ἱ) was aged. 
Transitivity: is used to check the possibility of indirect contact for nodes and is calculated based on Equation 3. For example, in Equation 3, “if node A frequently encounters node B, and node B frequently encounters node C, then node C probably is a good node to forward messages destined for node A”. 

P (A, C) = P (A, C) old + (1 – P (A, C) old) P(A, B) P(B, C) × …………….………………… (3)
For the design of the proposed energy-efficient PRoPHET routing protocol, the same assumptions that were utilized in the original routing protocols are considered, but the following energy-efficient parameters are also introduced:

Minimum energy threshold (T): This is a value set by the network administrator (in Joules) to prevent relay nodes that have low energy level from being selected as next-hop nodes. Before a source node can forward a message copy to an encountered relay node, the energy level of the neighbor node is checked so that if it is less than the prescribed value T, the message copy will not be sent to that node. This mechanism assists in selecting only the nodes that have enough energy to deliver the message copy to the destination node or closer to the destination node. However, if the encountered relay node is the destination node, the message copy is forwarded directly to it regardless of the energy level of that node. 
One-hop acknowledgment mechanism: This is implemented to remove the copies of an already delivered message from the network and to prevent unnecessary transmission of such message. In this mechanism, when the message copy has been delivered to its destination, the destination node sends an acknowledgment message to the last encounter node that has sent the message. Both nodes update their acknowledgment tables and drop the message from their buffer to prevent it from further relaying. After a message has been dropped by the two nodes, both nodes update the acknowledgment table and flood their updated acknowledgment table to the network so that any node that comes in contact with a node having updated acknowledgment table can update their own acknowledgment table and delete the already delivered messages from their buffer.